Add first version of CTaintASTConsumer
Add the CTaintASTConsumer.cc to perform taint analysis for a subset of
C89 programs. This analysis computes the labels, flows, and kill
functions for a C89 source file in this project's language subset. It
then computes the least solution for the program's gen function and
Entry and Exit equations. Currently we only print the Entry and Exit
results, but (currently commented out) code exists to print the results
of the other functions as well.